Added Top and Bottom Wall with Material. Made Prefab
So that the pixel will tile, the sprite needs to be set to Single with the Mesh Type of Full Rect
The material will color it's texture by Emission color.
It's important that the Width & Height be set to 1. And the Collider size also be set to 1x1. That way you can control it by scale.